pyproject.toml
[tool.poetry]
name = "cdk-example-app"
version = "0.1.0"
description = "AWS CDK Python experiments"
authors = ["houbie <ivo@houbrechts-it.be>"]

[tool.poetry.dependencies]
python = "^3.8"
aws-lambda-powertools = "^1.9.1"
pynamodb = "^5.0.0"
Flask = "^1.1.2"
connexion = "^2.7.0"
requests = "^2.25.1"
opentelemetry-sdk = "^1.1.0"
opentelemetry-propagator-b3 = "^1.1.0"
opentelemetry-instrumentation-flask = "^0.20b0"
opentelemetry-instrumentation-requests = "^0.20b0"
opentelemetry-instrumentation-botocore = "^0.20b0"
serverless-wsgi = "^1.7.7"

[tool.poetry.dev-dependencies]
boto3 = "^1.16.43"
pytest = "^6.2.1"
pytest-mock = "^3.5.1"
locust = "^1.4.1"
pytest-dynamodb = "^2.0.1"
"aws-cdk.core" = "^1.101.0"
"aws-cdk.aws-lambda" = "^1.101.0"
"aws-cdk.aws-apigateway" = "^1.101.0"
"aws-cdk.aws-s3-notifications" = "^1.102.0"
"aws-cdk.aws-lambda-event-sources" = "^1.102.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
